+++ R/lb-keogh.R
@@ -7,11 +7,11 @@
 #'
 #' @param x A time series (reference).
 #' @param y A time series with the same length as \code{x} (query).
-#' @param window.size Window size for envelope calculation. See details.
+#' @param window.size Window size for envelop calculation. See details.
 #' @param norm Vector norm. Either \code{"L1"} for Manhattan distance or \code{"L2"} for Euclidean.
-#' @param lower.env Optionally, a pre-computed lower envelope for \strong{\code{y}} can be provided
+#' @param lower.env Optionally, a pre-computed lower envelop for \strong{\code{y}} can be provided
 #'   (non-proxy version only).
-#' @param upper.env Optionally, a pre-computed upper envelope for \strong{\code{y}} can be provided
+#' @param upper.env Optionally, a pre-computed upper envelop for \strong{\code{y}} can be provided
 #'   (non-proxy version only).
 #' @param force.symmetry If \code{TRUE}, a second lower bound is calculated by swapping \code{x} and
 #'   \code{y}, and whichever result has a \emph{higher} distance value is returned. The proxy
@@ -32,8 +32,8 @@
 #' @return A list with:
 #' \itemize{
 #'   \item \code{d}: The lower bound of the DTW distance.
-#'   \item \code{upper.env}: The time series of \code{y}'s upper envelope.
-#'   \item \code{lower.env}: The time series of \code{y}'s lower envelope.
+#'   \item \code{upper.env}: The time series of \code{y}'s upper envelop.
+#'   \item \code{lower.env}: The time series of \code{y}'s lower envelop.
 #' }
 #'
 #' @note
@@ -105,7 +105,7 @@ lb_keogh <- function(x, y, window.size = NULL, norm = "L1",
 
     ## NOTE: the 'window.size' definition varies betwen dtw/call_envelop and runmin/max
     if (is.null(lower.env) && is.null(upper.env)) {
-        envelopes <- call_envelop(y, window.size)
+        envelopes <- compute_envelop(y, window.size)
         lower.env <- envelopes$lower
         upper.env <- envelopes$upper
 
@@ -117,10 +117,10 @@ lb_keogh <- function(x, y, window.size = NULL, norm = "L1",
     }
 
     if (length(lower.env) != length(x))
-        stop("Length mismatch between 'x' and the lower envelope")
+        stop("Length mismatch between 'x' and the lower envelop")
 
     if (length(upper.env) != length(x))
-        stop("Length mismatch between 'x' and the upper envelope")
+        stop("Length mismatch between 'x' and the upper envelop")
 
     D <- rep(0, length(x))
 
@@ -150,7 +150,7 @@ lb_keogh <- function(x, y, window.size = NULL, norm = "L1",
 }
 
 # ========================================================================================================
-# Loop without using native 'proxy' looping (to avoid multiple calculations of the envelope)
+# Loop without using native 'proxy' looping (to avoid multiple calculations of the envelop)
 # ========================================================================================================
 
 lb_keogh_proxy <- function(x, y = NULL, window.size = NULL, norm = "L1", ...,
@@ -180,7 +180,7 @@ lb_keogh_proxy <- function(x, y = NULL, window.size = NULL, norm = "L1", ...,
 
     retclass <- "crossdist"
 
-    envelops <- lapply(y, function(s) { call_envelop(s, window.size) })
+    envelops <- lapply(y, function(s) { compute_envelop(s, window.size) })
 
     lower.env <- lapply(envelops, "[[", "lower")
     upper.env <- lapply(envelops, "[[", "upper")