+++ R/utils.R
@@ -123,19 +123,6 @@ validate_pairwise <- function(x, y) {
 # Helper C/C++ functions
 # ========================================================================================================
 
-# Envelop calculation
-call_envelop <- function(series, window.size) {
-    check_consistency(series, "ts")
-    window.size <- check_consistency(window.size, "window")
-    window.size <- window.size * 2L + 1L
-
-    ## NOTE: window.size in this function is window.size*2 + 1, thus the 2L below
-    if (window.size > (2L * length(series)))
-        stop("Window cannot be greater or equal than the series' length.")
-
-    .Call("envelop", series, window.size, PACKAGE = "dtwclust")
-}
-
 # Create combinations of all possible pairs
 call_pairs <- function(n = 2L, lower = TRUE) {
     if (n < 2L) stop("At least two elements are needed to create pairs.")